create github client test utilities
